Added support for a file local-system-specific.pro in the source tree root folder (where the photivo.pro is located). The purpose of that file is to add special settings that are only valid for an individual computer.
For example on Windows Qt Creator tends to not recognize all relevant include paths of your
MinGW installation automatically, i.e. missing auto-complete and all the usual IDE goodies.

Because local-system-specific.pro holds individual configuration settings that are not
applicable to all toolchains and programming environments the file itself is not version
controlled. It is only imported by photivoProject.pro.