Prevent a divide by zero crash during ptDcRaw::identify() with certain jpeg files.
The JFIF specification requires that a valid JPEG/JFIF file starts with the SOI marker immediately
followed by an APP0 marker identifying the file as JFIF. However there are files that put exif
data immediately after the SOI marker and then the JFIF marker follows.
As far as I understand the specs such files are technically invalid, but they exist. They are
probably rare because the problem has been in the code for at least a year no one complained
earlier.