Sorted out mask types and overlay modes and the associated integer constants. At least partly replaced them with proper enums. More enums to follow when more of the affected filters will be ported to the new architecture.
Rant: If I see one more "short" abused as "bool" or one more block of "const short" abused
as a simulation of an enumeration I am going to explode! This disease is especially
wide-spread in Photivo's core image processing code. I can kind of understand where it
comes from. Things like that may be a valid programming style in C (I woudn't know, I'm
not a C programmer). In C++ this is just plain stupid. C++ fundamentally relies on its
rich type system. Circumventing it makes code error prone, hard to read and a nightmare to
maintain. Use the effing type system! That's what it's there for! [end of rant]